+++ tests/test_image.py
@@ -239,8 +239,8 @@ def test_Image_basic():
                 assert im2_cview.imag(x,y) == 0
 
                 value3 = 10*x + y
-                im1.addValue(x,y, value3-value2)
-                im2_view[x,y] += value3-value2
+                im1.addValue(x,y, np.int64(value3-value2))
+                im2_view[x,y] += np.int64(value3-value2)
                 assert im1[galsim.PositionI(x,y)] == value3
                 assert im1.view()[x,y] == value3
                 assert im1.view(make_const=True)[galsim.PositionI(x,y)] == value3
